Simplify proxying legacy funding finder
