+++ middleware/securityHeaders.test.js
@@ -14,14 +14,4 @@ describe('securityHeaders', () => {
         expect(res.get('Content-Security-Policy')).to.contain("default-src 'self'");
         expect(res.get('Content-Security-Policy')).to.exist;
     });
-
-    it('should not add Content-Security-Policy header for exempt URLs', () => {
-        const req = httpMocks.createRequest({
-            path: '/funding/funding-finder'
-        });
-        const res = httpMocks.createResponse();
-
-        defaultSecurityHeaders()(req, res, () => {});
-        expect(res.get('Content-Security-Policy')).to.not.exist;
-    });
 });
